+++ mobile/android/android-components/components/service/firefox-accounts/src/main/java/mozilla/components/service/fxa/sync/SyncManager.kt
@@ -107,6 +107,11 @@ abstract class SyncManager(
     // Currently the interfaces are the same, hence the name "pass-through".
     private val dispatcherStatusObserver = PassThroughSyncStatusObserver(syncStatusObserverRegistry)
 
+    /**
+     * Indicates if sync is currently running.
+     */
+    internal fun isSyncActive() = syncDispatcher?.isSyncActive() ?: false
+
     internal fun registerSyncStatusObserver(observer: SyncStatusObserver) {
         syncStatusObserverRegistry.register(observer)
     }